Saving...

+++ 2022/11/28/generate_int.cpp
@@ -4,7 +4,7 @@
 
 int main() {
   size_t N = 10000;
-  double *x = new double[N];
+  int64_t *x = new int64_t[N];
   std::random_device rd;
   std::default_random_engine eng(rd());
   std::uniform_int_distribution<int64_t> distr(-127,127);
